add: 함수별 파일 분리, 구조변경
