+++ NumberBaseball/NumberBaseball/CheckStrikeAndBall.swift
@@ -1,0 +1,25 @@
+//
+//  CheckWinOrLose.swift
+//  NumberBaseball
+//
+//  Created by 김민제 on 1/4/24.
+//
+
+import Foundation
+
+extension NumberBaseBall {
+    func compare(comNum: [String], userNum: [String]){
+        var userNumList: Array<String> = []
+        
+        guard let inputUser = readLine() else {
+            return
+        }
+        
+        var strike: Int = 0
+        var ball: Int = 0
+        
+        
+    }
+
+}
+
