+++ NumberBaseball/NumberBaseball/UserInputNumber.swift
@@ -1,0 +1,12 @@
+//
+//  UserInputNumber.swift
+//  NumberBaseball
+//
+//  Created by 김민제 on 1/4/24.
+//
+
+import Foundation
+
+extension NumberBaseBall {
+    
+}
